moved fixtures to test_fixtures:
